Reorganize UI elements into separate packages
